+++ src/family/src/application.cpp
@@ -1,3 +1,6 @@
+#include <QDir>
+#include <QFileInfo>
+
 #include <rgl_message_box.h>
 
 #include "application.h"
@@ -42,6 +45,7 @@ void Application::initialize()
     this->cloudConnectionHandler->setMessageBoxParentWidget(this->mainWindow);
     this->mainWindow->show();
 
+    QObject::connect(this->session,&Session::treeFileLoaded,this,&Application::onTreeFileLoaded);
     QObject::connect(this->session,&Session::treeFileChanged,this,&Application::onTreeFileChanged);
     QObject::connect(this->session,&Session::treeFileRemoved,this,&Application::onTreeFileRemoved);
 }
@@ -49,6 +53,55 @@ void Application::initialize()
 void Application::finalize()
 {
 
+}
+
+//! Check whether given file is located at or below given directory.
+static bool isBelowDirectory(const QString &fileName, const QString &directory)
+{
+    if (fileName.isEmpty() || directory.isEmpty())
+    {
+        return false;
+    }
+
+    QString relativePath = QDir(directory).relativeFilePath(QFileInfo(fileName).absolutePath());
+
+    return !(relativePath.isEmpty() ||
+             QDir::isAbsolutePath(relativePath) ||
+             relativePath == ".." ||
+             relativePath.startsWith("../"));
+}
+
+void Application::onTreeFileLoaded(const QString &fileName)
+{
+    QString dataDir = this->applicationSettings->getDataDir();
+
+    if (isBelowDirectory(fileName,dataDir))
+    {
+        return;
+    }
+
+    QString question = tr("The family tree file is not stored in the data directory.")
+                     + "<pre>" + fileName + "</pre>"
+                     + tr("Would you like to move it to the data directory?")
+                     + "<pre>" + dataDir + "</pre>";
+
+    if (RMessageBox::question(this->mainWindow,tr("Tree file outside of the data directory"),question) != RMessageBox::Yes)
+    {
+        return;
+    }
+
+    QString newFileName = QDir(dataDir).absoluteFilePath(QFileInfo(fileName).fileName());
+
+    if (QFileInfo::exists(newFileName))
+    {
+        QString message = tr("The data directory already contains a file with the same name.")
+                        + "<pre>" + newFileName + "</pre>"
+                        + tr("The family tree file was not moved.");
+        RMessageBox::warning(this->mainWindow,tr("Tree file was not moved"),message);
+        return;
+    }
+
+    this->session->moveTreeFile(newFileName);
 }
 
 void Application::onTreeFileChanged(const QString &fileName)